SessionPills file rename + hook
